+++ src/model/database/models/Rol.js
@@ -1,15 +1,15 @@
-module.exports = (sequelize, DataTypes) =>{
+module.exports = (sequelize, DataTypes) => {
     let alias = 'Rol';
     let columns = {
         id_rol : {
-            type: DataTypes.INTEGER,
+            type: DataTypes.BIGINT,
             allowNull: false,
             primaryKey: true,
             autoIncrement: true
         },
         name: {
             type: DataTypes.STRING,
-            allowNull:false
+            allowNull: false
         }
     }
     let config = {
